# web/docs/swagger.yaml
definitions:
...
  models.ResIpStaticLogModel:
    properties:
      account:
        type: string
      city:
        type: string
      country:
        type: string
      create_time:
        type: string
      expire_time:
        type: string
      forward:
        type: string
      id:
        type: integer
      ip:
        type: string
      is_expire:
        type: integer
      password:
        type: string
      port:
        type: integer
      remark:
        type: string
      state:
        type: string
    type: object
...
  models.ResStaticRecordModel:
    properties:
      city:
        type: string
      country:
        type: string
      create_time:
        type: string
      expire_time:
        type: string
      id:
        type: integer
      ip:
        type: string
      is_expire:
        type: integer
      port:
        type: integer
      remark:
        type: string
      state:
        type: string
    type: object
...
host: textapi.360proxy.net
info:
  contact: {}
  description: This 360 API
  title: 360 API
  version: "1.0"
paths:
...
  /web/score/exchange:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 积分兑换流量
      parameters:
      - description: 用户登录信息
        in: formData
        name: session
        required: true
        type: string
      - description: 兑换流量数量
        in: formData
        name: flow
        required: true
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: all_score：总积分，use_score：已使用积分，score：兑换后剩余积分
          schema:
            additionalProperties: true
            type: object
      summary: 积分兑换流量
      tags:
      - 个人中心 - 积分
  /web/score/exchange_flow_day:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 积分兑换不限量流量
      parameters:
      - description: 用户登录信息
        in: formData
        name: session
        required: true
        type: string
      - description: 兑换流量天数
        in: formData
        name: flow
        required: true
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: all_score：总积分，use_score：已使用积分，score：兑换后剩余积分
          schema:
            additionalProperties: true
            type: object
      summary: 积分兑换不限量流量
      tags:
      - 个人中心 - 积分
  /web/score/feedback:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 反馈上报信息
      parameters:
      - description: 用户登录信息
        in: formData
        name: session
        required: true
        type: string
      - description: 反馈内容
        in: formData
        name: content
        required: true
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: 成功
          schema:
            type: object
      summary: 反馈上报信息
      tags:
      - 个人中心 - 积分
  /web/score/free:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 免费获取积分上报信息
      parameters:
      - description: 用户登录信息
        in: formData
        name: session
        required: true
        type: string
      - description: 语言
        in: formData
        name: lang
        required: true
        type: string
      - description: 邮箱
        in: formData
        name: email
        required: true
        type: string
      - description: 手机号
        in: formData
        name: phone
        required: true
        type: string
      - description: 公司名称
        in: formData
        name: company
        required: true
        type: string
      - description: 公司地址
        in: formData
        name: address
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: 成功
          schema:
            type: object
      summary: 免费获取积分上报信息
      tags:
      - 个人中心 - 积分
  /web/score/info:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 获取积分信息
      parameters:
      - description: 用户登录信息
        in: formData
        name: session
        required: true
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: country：国家，all_score：获取的总积分，use_score：已使用积分，score：剩余积分，expire：过期时间，feedback：反馈弹窗的状态
            0无反馈 1待处理 2成功  3失败未填写反馈 4 失败已填写反馈
          schema:
            additionalProperties: true
            type: object
      summary: 获取积分信息
      tags:
      - 个人中心 - 积分
  /web/score/record:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 获取积分记录
      parameters:
      - description: 用户登录信息
        in: formData
        name: session
        required: true
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: score：积分记录（值为[]ResLogUserScore{}模型），use：使用记录（值为[]ResLogUserScore{}模型）
          schema:
            additionalProperties: true
            type: object
      summary: 获取积分记录
      tags:
      - 个人中心 - 积分
...
  /web/static/use_download:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 使用记录下载
      parameters:
      - description: 用户登录信息
        in: formData
        name: session
        required: true
        type: string
      - description: 开始日期
        in: formData
        name: start_date
        type: string
      - description: 结束日期
        in: formData
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: 成功
          schema:
            type: object
      summary: 使用记录下载
      tags:
      - 个人中心 - 静态住宅代理
  /web/static/use_list:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 静态 IP使用列表
      parameters:
      - description: 用户登录信息
        in: formData
        name: session
        required: true
        type: string
      - description: ip /备注筛选
        in: formData
        name: ip
        type: string
      - description: 状态
        in: formData
        name: status
        type: string
      - description: 排序字段
        in: formData
        name: field
        type: string
      - description: 排序  1降序
        in: formData
        name: sorter
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: 成功
          schema:
            items:
              $ref: '#/definitions/models.ResIpStaticLogModel'
            type: array
      summary: 静态 IP使用列表
      tags:
      - 个人中心 - 静态住宅代理
  /web/static/use_record:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 静态IP  提取记录
      parameters:
      - description: 用户登录信息
        in: formData
        name: session
        required: true
        type: string
      - description: 开始日期
        in: formData
        name: start_date
        type: string
      - description: 结束日期
        in: formData
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: 成功
          schema:
            items:
              $ref: '#/definitions/models.ResStaticRecordModel'
            type: array
      summary: 静态IP  提取记录
      tags:
      - 个人中心 - 静态住宅代理
...
schemes:
- http
- https
swagger: "2.0"